+++ divik/cluster/_divik/__init__.py
@@ -1,3 +1,3 @@
 """DiviK algorithm implementation."""
 
-from ._sklearn import DiviK
+from ._dunn_sklearn import DiviK
